pyproject.toml
[project]
authors = [{ name = "Lucas", email = "llevesque@lanets.ca" }]
name = "label_generator"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"
dependencies = ["Pillow"]

[tool.black]
line-length = 88
